pyproject.toml
[project]
name = "hsipl_autotool"
version = "0.1.0"
description = "Hyperspectral Image Processing Auto Tools Server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [    
    "numpy",
    "matplotlib",
    "fastmcp>=2.2.5",
    ]
[project.scripts]
hsipl-autotool = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hsipl_autotool"]

[tool.mcp]
module = "server"         # 指向 server.py
server_function = "mcp"   # 指向 mcp = FastMCP(...)
timeout = 300             # 超時設定

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.389",
    "ruff>=0.7.3",
]
